Move statements to configTools
